+++ src/track/track_utils.h
@@ -66,7 +66,7 @@ double tracker_get_lock_time(const tracker_time_info_t *time_info,
 u16 tracker_load_cc_data(tracker_cc_data_t *cc_data);
 
 void tracker_set_carrier_phase_offset(const tracker_info_t *info,
-                                      double carrier_phase_offset);
+                                      s64 carrier_phase_offset);
 void tracker_carrier_phase_offsets_adjust(double dt);
 
 tracker_t *tracker_channel_get_by_mesid(const me_gnss_signal_t mesid);
